+++ core-ui/src/shared/components/CreateModal/CollapsibleSection/CollapsibleSection.js
@@ -2,31 +2,21 @@ import { Icon } from 'fundamental-react';
 import React from 'react';
 import './CollapsibleSection.scss';
 
-export function CollapsibleSection({
-  defaultOpen = false,
-  title,
-  actions,
-  children,
-}) {
+export function CollapsibleSection({ defaultOpen, title, actions, children }) {
   const [open, setOpen] = React.useState(defaultOpen);
   const actionsRef = React.useRef();
+  const iconGlyph = open ? 'navigation-down-arrow' : 'navigation-right-arrow';
 
   const toggle = e => {
     // ignore events from actions
-    if (actionsRef.current?.contains(e.target)) {
-      return;
-    }
-    setOpen(!open);
+    if (!actionsRef.current?.contains(e.target)) setOpen(!open);
   };
 
   return (
     <div className="create-modal__collapsible-section">
       <header onClick={toggle}>
         <div>
-          <Icon
-            ariaHidden
-            glyph={open ? 'navigation-down-arrow' : 'navigation-right-arrow'}
-          />
+          <Icon ariaHidden glyph={iconGlyph} />
           {title}
         </div>
         <div ref={actionsRef}>{actions}</div>
