+++ core-ui/src/shared/components/CreateModal/CreateModal.scss
