+++ core-ui/src/shared/components/CreateModal/ModeSelector/ModeSelector.js
@@ -2,15 +2,11 @@ import React from 'react';
 import { Button, ButtonSegmented } from 'fundamental-react';
 import './ModeSelector.scss';
 
-export const MODE_SIMPLE = 'MODE_SIMPLE';
-export const MODE_ADVANCED = 'MODE_ADVANCED';
-export const MODE_YAML = 'MODE_YAML';
-
 export function ModeSelector({ mode, setMode }) {
   const modeButtons = [
-    { mode: MODE_SIMPLE, label: 'Simple' },
-    { mode: MODE_ADVANCED, label: 'Advanced' },
-    { mode: MODE_YAML, label: 'YAML' },
+    { mode: ModeSelector.MODE_SIMPLE, label: 'Simple' },
+    { mode: ModeSelector.MODE_ADVANCED, label: 'Advanced' },
+    { mode: ModeSelector.MODE_YAML, label: 'YAML' },
   ];
 
   return (
@@ -30,3 +26,7 @@ export function ModeSelector({ mode, setMode }) {
     </div>
   );
 }
+
+ModeSelector.MODE_SIMPLE = 'MODE_SIMPLE';
+ModeSelector.MODE_ADVANCED = 'MODE_ADVANCED';
+ModeSelector.MODE_YAML = 'MODE_YAML';
