+++ resources/web/deployment.yaml
@@ -16,7 +16,7 @@ spec:
     spec:
       containers:
         - name: busola
-          image: eu.gcr.io/kyma-project/busola-web:PR-335
+          image: eu.gcr.io/kyma-project/busola-web:PR-345
           imagePullPolicy: Always
           resources:
             requests:
